buildmate/requirements.txt
pytest
requests
beautifulsoup4
flask
google-generativeai